requirements.txt
jupyterlab
Pillow
opencv-python
matplotlib
numpy
torch
torchvision
pandas